+++ src/librustc_passes/lib.rs
@@ -32,6 +32,7 @@ mod liveness;
 pub mod loops;
 mod reachable;
 mod region;
+pub mod stability;
 
 pub fn provide(providers: &mut Providers<'_>) {
     check_const::provide(providers);
@@ -43,4 +44,5 @@ pub fn provide(providers: &mut Providers<'_>) {
     intrinsicck::provide(providers);
     reachable::provide(providers);
     region::provide(providers);
+    stability::provide(providers);
 }
